requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
mlflow==2.13.1
datasets==2.19.1
joblib==1.4.2
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1
streamlit==1.34.0
PyYAML==6.0.1
